BACKLOG-10738: Some refactoring
- removed dependency on @material-ui/styles 4.x
- use DOM reference of the drawer instead of handlebar
- moved ResizingHandlerBar styling to the component
