+++ package.json
@@ -31,7 +31,6 @@
     "@material-ui/core": "^3.6.2",
     "@material-ui/icons": "^3.0.2",
     "@material-ui/lab": "^3.0.0-alpha.16 ",
-    "@material-ui/styles": "^4.2.1",
     "classnames": "^2.2.5",
     "connected-react-router": "^4.4.1",
     "copy-to-clipboard": "^3.0.8",
