+++ src/javascript/ContentManager/ContentRoute/ContentLayout/ResizableDrawer/ResizableDrawer.jsx
@@ -1,30 +1,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {Drawer, withStyles} from '@material-ui/core';
+import {Drawer, RootRef} from '@material-ui/core';
 import ResizingHandleBar from './ResizingHandleBar';
-
-const constants = Object.freeze({
-    handleBarOffset: 4
-});
-
-const styles = theme => ({
-    handleBar: {
-        backgroundColor: 'transparent',
-        cursor: 'col-resize',
-        minHeight: '100%',
-        height: '100%',
-        position: 'absolute',
-        right: `-${constants.handleBarOffset}px`,
-        width: `calc(${constants.handleBarOffset}px * 2 + 1px)`,
-        zIndex: theme.zIndex.drawer
-    }
-});
 
 export class ResizableDrawer extends React.Component {
     constructor(props) {
         super(props);
         this.state = {resizing: false};
-        this.handleBar = React.createRef();
+        this.drawer = React.createRef();
 
         this.startResizing = this.startResizing.bind(this);
         this.stopResizing = this.stopResizing.bind(this);
@@ -38,7 +21,7 @@ export class ResizableDrawer extends React.Component {
 
         this.setState({resizing: true});
 
-        const {ownerDocument} = this.handleBar.current;
+        const {ownerDocument} = this.drawer.current;
         ownerDocument.addEventListener('mousemove', this.resize);
         ownerDocument.addEventListener('mouseup', this.stopResizing);
     }
@@ -50,7 +33,7 @@ export class ResizableDrawer extends React.Component {
 
         this.setState({resizing: false});
 
-        const {ownerDocument} = this.handleBar.current;
+        const {ownerDocument} = this.drawer.current;
         ownerDocument.removeEventListener('mousemove', this.resize);
         ownerDocument.removeEventListener('mouseup', this.stopResizing);
     }
@@ -62,7 +45,7 @@ export class ResizableDrawer extends React.Component {
 
         event.preventDefault();
 
-        const initialPosition = this.handleBar.current.getBoundingClientRect().right - constants.handleBarOffset;
+        const initialPosition = this.drawer.current.getBoundingClientRect().right;
         const offset = event.pageX - initialPosition;
         const newWidth = this.props.width + offset;
 
@@ -70,15 +53,16 @@ export class ResizableDrawer extends React.Component {
     }
 
     render() {
-        const {children, classes, width, open, onResized, ...otherProps} = this.props;
-        const {handleBar: handleBarClass, ...otherClasses} = classes;
+        const {children, onResized, width, open, ...otherProps} = this.props;
         return (
             <>
                 {open &&
-                <Drawer width={width + 'px'} classes={otherClasses} open={open} {...otherProps}>
-                    <ResizingHandleBar ref={this.handleBar} className={handleBarClass} onMouseDown={this.startResizing}/>
-                    {children}
-                </Drawer>
+                    <RootRef rootRef={this.drawer}>
+                        <Drawer width={width + 'px'} open={open} {...otherProps}>
+                            <ResizingHandleBar onMouseDown={this.startResizing}/>
+                            {children}
+                        </Drawer>
+                    </RootRef>
                 }
             </>
         );
@@ -87,10 +71,9 @@ export class ResizableDrawer extends React.Component {
 
 ResizableDrawer.propTypes = {
     children: PropTypes.node,
-    classes: PropTypes.object.isRequired,
     onResized: PropTypes.func.isRequired,
     open: PropTypes.bool.isRequired,
     width: PropTypes.number.isRequired
 };
 
-export default withStyles(styles)(ResizableDrawer);
+export default ResizableDrawer;
